+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Target Scheduler
 
-## 4.7.6.0 - 2024-09-03
+## 4.7.6.1 - 2024-09-06
 * Don't abort a flat exposure if setting flat panel brightness throws an error, just warn
 * Changed the time at which moon avoidance is evaluated, now halfway through target's minimum time
+* Removed rotation as one of the comparison criteria for flats
 
 ## 4.7.5.0 - 2024-09-01
 * Added ability to change scheduler preview start time to now.